music_player refactored
