+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:theme="@style/Theme.AtmConsultancy" />
 
         <activity
-            android:name=".feature.media_player.MediaPlayerActivity"
+            android:name=".feature.media_player.activities.MediaPlayerActivity"
             android:label="@string/media_player"
             android:theme="@style/Theme.AndroidUdemyCourse" />
     </application>
