+++ app/src/main/res/layout/activity_media_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.media_player.MediaPlayerActivity">
+    tools:context=".feature.media_player.activities.MediaPlayerActivity">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/playMusicTextView"
