+++ app/src/main/res/layout/fragment_music_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.media_player.fragments.MusicPlayerFragments">
+    tools:context=".feature.media_player.fragments.MusicPlayerFragment">
 
     <androidx.appcompat.widget.AppCompatSeekBar
         android:id="@+id/volumeSeekBar"
